Move ContentWrapper to correct directory
