+++ src/app/containers/App/App.js
@@ -12,7 +12,7 @@ import CreateWallet from '../CreateWallet'
 import ImportWallet from '../ImportWallet'
 import ExportWallet from '../ExportWallet'
 import Config from '../Config'
-import ContentWrapper from '../ContentWrapper'
+import ContentWrapper from '../../components/ContentWrapper'
 import StartPage from '../../components/StartPage'
 import Login from '../../components/Login'
 import Header from '../../components/Header'
